+++ CHANGELOG.md
@@ -17,7 +17,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Frame2D/3D: fix rotation
 
 ### Refactor
-- Circle2D
+- Circle2D: Now, it needs a Frame2D and a radius instead of a Center and a Radius. This allows to easily control the circle's direction (clockwise/counterclockwise)
+- Arc2D: Arc 2D now must follow the same rotation direction of its circle.
 
 ### Changed
 -